requirements.txt
ansicon==1.89.0
async-generator==1.10
attrs==22.1.0
blessed==1.19.1
Brotli==1.0.9
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
ConfigArgParse==1.5.3
Flask==2.2.2
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
gevent==21.12.0
geventhttpclient==2.0.2
greenlet==1.1.3
h11==0.13.0
idna==3.4
inquirer==2.9.2
inquirerpy==0.3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jinxed==1.2.0
joblib==1.2.0
locust==2.12.1
MarkupSafe==2.1.1
MouseInfo==0.1.3
msgpack==1.0.4
nltk==3.8.1
outcome==1.2.0
pfzy==0.3.4
prompt-toolkit==3.0.38
psutil==5.9.2
PyAutoGUI==0.9.53
pycparser==2.21
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
python-editor==1.0.4
pytweening==1.0.4
pywin32==304
PyYAML==6.0
pyzmq==24.0.1
readchar==3.0.5
regex==2023.3.23
requests==2.28.1
roundrobin==0.0.4
selenium==4.9.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tqdm==4.65.0
trio==0.21.0
trio-websocket==0.9.2
typing_extensions==4.3.0
urllib3==1.26.12
wcwidth==0.2.5
Werkzeug==2.2.2
wsproto==1.2.0
zope.event==4.5.0
zope.interface==5.4.0
